% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pizza_topping_prices.R
\name{pizza_topping_prices}
\alias{pizza_topping_prices}
\title{Calculating Pizza Topping Prices}
\usage{
pizza_topping_prices(pineapples, onions, olives)
}
\arguments{
\item{pineapples}{number of the order of the best pizza topping}

\item{onions}{number of order of red onions}

\item{olives}{number of order of black olives}
}
\value{
the total price of the pizza toppings
}
\description{
Calculating Pizza Topping Prices
}
\examples{
pizza_topping_prices(pineapples = 2, onions = 4, olives = 1)
}
